refactor: изменена файловая структура проекта
