+++ src/components/modal.js
@@ -1,5 +1,26 @@
-export { openPopup, closePopup, isLoading };
-import { popupOpened, popupClose } from './variables.js';
+import { popupOpened, popupClose } from '../utils/constants.js';
+
+export function openPopup(popup) {
+  popup.classList.add('popup_opened');
+  document.addEventListener('keydown', closeEscape);
+  popup.addEventListener('click', closeClick);
+};
+
+export function closePopup(popup) {
+  popup.classList.remove('popup_opened');
+  document.removeEventListener('keydown', closeEscape);
+  popup.removeEventListener('click', closeClick);
+};
+
+export function isLoading(poupForm, submitButtonSelector, loading) {
+  const button = poupForm.querySelector(submitButtonSelector);
+  if(loading) {
+    button.textContent = 'Сохранение...'
+  }
+  else {
+    button.textContent = 'Сохранить'
+  }
+}
 
 function closeEscape(event) {
   if (event.key === 'Escape') {
@@ -10,27 +31,5 @@ function closeEscape(event) {
 function closeClick(event) {
   if(event.target.classList.contains(popupOpened) || event.target.classList.contains(popupClose)) {
     closePopup(event.currentTarget);
-  }
-}
-
-function openPopup(popup) {
-  popup.classList.add('popup_opened');
-  document.addEventListener('keydown', closeEscape);
-  popup.addEventListener('click', closeClick);
-};
-
-function closePopup(popup) {
-  popup.classList.remove('popup_opened');
-  document.removeEventListener('keydown', closeEscape);
-  popup.removeEventListener('click', closeClick);
-};
-
-function isLoading(poupForm, submitButtonSelector, loading) {
-  const button = poupForm.querySelector(submitButtonSelector);
-  if(loading) {
-    button.textContent = 'Сохранение...'
-  }
-  else {
-    button.textContent = 'Сохранить'
   }
 }
