+++ src/components/validate.js
@@ -1,18 +1,11 @@
-export { enableValidation, checkInputValidity };
-
-function showInputError(formElement, inputElement, errorMessage, inputErrorClass) {
-  const errorElement = formElement.querySelector(`.${inputElement.id}-error`);
-  errorElement.textContent = errorMessage;
-  inputElement.classList.add(inputErrorClass);
+export function enableValidation(config) {
+  const formList = Array.from(document.querySelectorAll(config.formSelector));
+  formList.forEach((formElement) => {
+    setEventListeners(formElement, config.inputSelector, config.submitButtonSelector, config.inputErrorClass);
+  });
 };
 
-function hideInputError(formElement, inputElement, inputErrorClass) {
-  const errorElement = formElement.querySelector(`.${inputElement.id}-error`);
-  errorElement.textContent = '';
-  inputElement.classList.remove(inputErrorClass);
-};
-
-function checkInputValidity(formElement, inputElement, inputErrorClass) {
+export function checkInputValidity(formElement, inputElement, inputErrorClass) {
   if (inputElement.validity.patternMismatch) {
     inputElement.setCustomValidity(inputElement.dataset.errorMessage);
   } 
@@ -27,6 +20,18 @@ function checkInputValidity(formElement, inputElement, inputErrorClass) {
   };
 };
 
+function showInputError(formElement, inputElement, errorMessage, inputErrorClass) {
+  const errorElement = formElement.querySelector(`.${inputElement.id}-error`);
+  errorElement.textContent = errorMessage;
+  inputElement.classList.add(inputErrorClass);
+};
+
+function hideInputError(formElement, inputElement, inputErrorClass) {
+  const errorElement = formElement.querySelector(`.${inputElement.id}-error`);
+  errorElement.textContent = '';
+  inputElement.classList.remove(inputErrorClass);
+};
+
 function setEventListeners(formElement, popupInput, submitButtonSelector, inputErrorClass) {
   const inputList = Array.from(formElement.querySelectorAll(popupInput));
   const buttonElement = formElement.querySelector(submitButtonSelector);
@@ -36,13 +41,6 @@ function setEventListeners(formElement, popupInput, submitButtonSelector, inputE
       checkInputValidity(formElement, inputElement, inputErrorClass);
       toggleButtonState(inputList, buttonElement);
     });
-  });
-};
-
-function enableValidation(config) {
-  const formList = Array.from(document.querySelectorAll(config.formSelector));
-  formList.forEach((formElement) => {
-    setEventListeners(formElement, config.inputSelector, config.submitButtonSelector, config.inputErrorClass);
   });
 };
